requirements.txt
win11toast>=0.3.0
watchdog>=3.0.0
pyperclip>=1.8.2
pystray>=0.19.0
pillow>=9.0.0